Phase 4 SVG: split SvgRasterizer into feature collaborators (no behavior change)
Addresses Roland's repeated recommendation (before SVG part 7): the 1435-line
SvgRasterizer carried filters, markers, masks, clip-path, patterns, style cascade,
and geometry. Extracted the four self-contained feature areas into internal
collaborators, leaving SvgRasterizer (now 832 lines) as the render dispatch +
shapes + style cascade + shared low-level helpers:

- SvgFilters    — <filter> image-filter chain (feGaussianBlur/feOffset/
                  feDropShadow/feColorMatrix + routing/region flagging).
- SvgMarkers    — marker-start/-mid/-end vertex placement + orient + viewport.
- SvgClipMask   — clip-path + mask element refs + the objectBoundingBox bbox.
- SvgPatterns   — <pattern> tile rendering → repeating shader.

The shared low-level helpers (RenderElement, BuildShapePath, ResolveStyle, Attr,
Num, Len/LenAxis, TrimUnit, ParseViewBox, ParsePoints, HasAnyAttr, PaintServerId,
MaxDepth) are now internal so the collaborators call back into SvgRasterizer.
Byte-identical: all 169 SVG tests + the 6 byte-identity gates + AOT 2942DD1E
unchanged.
